+++ Nos/Views/Discover/DiscoverContentsView.swift
@@ -3,7 +3,7 @@ import SwiftUI
 import Dependencies
 
 struct DiscoverContentsView: View {
-    @ObservedObject var searchController: SearchController
+    private var searchController: SearchController
 
     @EnvironmentObject private var router: Router