+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added the Account Success onboarding screen. Currently behind the “New Onboarding Flow” feature flag. [#1599](https://github.com/planetary-social/nos/issues/1599)
 - Updated the Age Verification onboarding screen. Currently behind the “New Onboarding Flow” feature flag. [#1651](https://github.com/planetary-social/nos/issues/1651)
 - Track opening mentions with Posthog. [#1480](https://github.com/planetary-social/nos/issues/1480)
+- More ObservableObject to @Observable migrations [#1458](https://github.com/planetary-social/nos/issues/1458)
 - Avoid crash and print extra debugging details when a reposted note that has not finished loading is clicked. [#1669](https://github.com/planetary-social/nos/issues/1669)
 - Changed the term "mute" to "block". [#1681](https://github.com/planetary-social/nos/pull/1681)